==> scripts/02_data_merge.Rmd <==
---
title: "Data Merging"
author: "Allie Caughman"
date: "2023-01-02"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(tidyverse)
```

### Read in CSV

```{r}
# load all csvs, assigning the correct evolution scenario, climate scenario, and MPA size listed (see model_list.xlsx for details)

df1 <- read_csv(here::here("generated_data", "summary_data", "summary_m1.csv")) %>%
  mutate(evolution = "Yes") %>%
  mutate(mpa_size = "1 x 1") %>%
  mutate(climate = "Null")
df2 <- read_csv(here::here("generated_data", "summary_data", "summary_m2.csv")) %>%
  mutate(evolution = "No") %>%
  mutate(mpa_size = "1 x 1") %>%
  mutate(climate = "Null")
df3 <- read_csv(here::here("generated_data", "summary_data", "summary_m3.csv")) %>%
  mutate(evolution = "Yes") %>%
  mutate(mpa_size = "3 x 3") %>%
  mutate(climate = "Null")
df4 <- read_csv(here::here("generated_data", "summary_data", "summary_m4.csv")) %>%
  mutate(evolution = "No") %>%
  mutate(mpa_size = "3 x 3") %>%
  mutate(climate = "Null")
df5 <- read_csv(here::here("generated_data", "summary_data", "summary_m5.csv")) %>%
  mutate(evolution = "Yes") %>%
  mutate(mpa_size = "6 x 6") %>%
  mutate(climate = "Null")
df6 <- read_csv(here::here("generated_data", "summary_data", "summary_m6.csv")) %>%
  mutate(evolution = "No") %>%
  mutate(mpa_size = "6 x 6") %>%
  mutate(climate = "Null")
df7 <- read_csv(here::here("generated_data", "summary_data", "summary_m7.csv")) %>%
  mutate(evolution = "Yes") %>%
  mutate(mpa_size = "10 x 10") %>%
  mutate(climate = "Null")
df8 <- read_csv(here::here("generated_data", "summary_data", "summary_m8.csv")) %>%
  mutate(evolution = "No") %>%
  mutate(mpa_size = "10 x 10") %>%
  mutate(climate = "Null")

df9 <- read_csv(here::here("generated_data", "summary_data", "summary_m9.csv")) %>%
  mutate(evolution = "Yes") %>%
  mutate(mpa_size = "1 x 1") %>%
  mutate(climate = "Mean Shift")
df10 <- read_csv(here::here("generated_data", "summary_data", "summary_m10.csv")) %>%
  mutate(evolution = "No") %>%
  mutate(mpa_size = "1 x 1") %>%
  mutate(climate = "Mean Shift")
df11 <- read_csv(here::here("generated_data", "summary_data", "summary_m11.csv")) %>%
  mutate(evolution = "Yes") %>%
  mutate(mpa_size = "3 x 3") %>%
  mutate(climate = "Mean Shift")
df12 <- read_csv(here::here("generated_data", "summary_data", "summary_m12.csv")) %>%
  mutate(evolution = "No") %>%
  mutate(mpa_size = "3 x 3") %>%
  mutate(climate = "Mean Shift")
df13 <- read_csv(here::here("generated_data", "summary_data", "summary_m13.csv")) %>%
  mutate(evolution = "Yes") %>%
  mutate(mpa_size = "6 x 6") %>%
  mutate(climate = "Mean Shift")
df14 <- read_csv(here::here("generated_data", "summary_data", "summary_m14.csv")) %>%
  mutate(evolution = "No") %>%
  mutate(mpa_size = "6 x 6") %>%
  mutate(climate = "Mean Shift")
df15 <- read_csv(here::here("generated_data", "summary_data", "summary_m15.csv")) %>%
  mutate(evolution = "Yes") %>%
  mutate(mpa_size = "10 x 10") %>%
  mutate(climate = "Mean Shift")
df16 <- read_csv(here::here("generated_data", "summary_data", "summary_m16.csv")) %>%
  mutate(evolution = "No") %>%
  mutate(mpa_size = "10 x 10") %>%
  mutate(climate = "Mean Shift")

df17 <- read_csv(here::here("generated_data", "summary_data", "summary_m17.csv")) %>%
  mutate(evolution = "Yes") %>%
  mutate(mpa_size = "1 x 1") %>%
  mutate(climate = "El Nino/La Nina")
df18 <- read_csv(here::here("generated_data", "summary_data", "summary_m18.csv")) %>%
  mutate(evolution = "No") %>%
  mutate(mpa_size = "1 x 1") %>%
  mutate(climate = "El Nino/La Nina")
df19 <- read_csv(here::here("generated_data", "summary_data", "summary_m19.csv")) %>%
  mutate(evolution = "Yes") %>%
  mutate(mpa_size = "3 x 3") %>%
  mutate(climate = "El Nino/La Nina")
df20 <- read_csv(here::here("generated_data", "summary_data", "summary_m20.csv")) %>%
  mutate(evolution = "No") %>%
  mutate(mpa_size = "3 x 3") %>%
  mutate(climate = "El Nino/La Nina")
df21 <- read_csv(here::here("generated_data", "summary_data", "summary_m21.csv")) %>%
  mutate(evolution = "Yes") %>%
  mutate(mpa_size = "6 x 6") %>%
  mutate(climate = "El Nino/La Nina")
df22 <- read_csv(here::here("generated_data", "summary_data", "summary_m22.csv")) %>%
  mutate(evolution = "No") %>%
  mutate(mpa_size = "6 x 6") %>%
  mutate(climate = "El Nino/La Nina")
df23 <- read_csv(here::here("generated_data", "summary_data", "summary_m23.csv")) %>%
  mutate(evolution = "Yes") %>%
  mutate(mpa_size = "10 x 10") %>%
  mutate(climate = "El Nino/La Nina")
df24 <- read_csv(here::here("generated_data", "summary_data", "summary_m24.csv")) %>%
  mutate(evolution = "No") %>%
  mutate(mpa_size = "10 x 10") %>%
  mutate(climate = "El Nino/La Nina")

df25 <- read_csv(here::here("generated_data", "summary_data", "summary_m25.csv")) %>%
  mutate(evolution = "Yes") %>%
  mutate(mpa_size = "1 x 1") %>%
  mutate(climate = "Shocks")
df26 <- read_csv(here::here("generated_data", "summary_data", "summary_m26.csv")) %>%
  mutate(evolution = "No") %>%
  mutate(mpa_size = "1 x 1") %>%
  mutate(climate = "Shocks")
df27 <- read_csv(here::here("generated_data", "summary_data", "summary_m27.csv")) %>%
  mutate(evolution = "Yes") %>%
  mutate(mpa_size = "3 x 3") %>%
  mutate(climate = "Shocks")
df28 <- read_csv(here::here("generated_data", "summary_data", "summary_m28.csv")) %>%
  mutate(evolution = "No") %>%
  mutate(mpa_size = "3 x 3") %>%
  mutate(climate = "Shocks")
df29 <- read_csv(here::here("generated_data", "summary_data", "summary_m29.csv")) %>%
  mutate(evolution = "Yes") %>%
  mutate(mpa_size = "6 x 6") %>%
  mutate(climate = "Shocks")
df30 <- read_csv(here::here("generated_data", "summary_data", "summary_m30.csv")) %>%
  mutate(evolution = "No") %>%
  mutate(mpa_size = "6 x 6") %>%
  mutate(climate = "Shocks")
df31 <- read_csv(here::here("generated_data", "summary_data", "summary_m31.csv")) %>%
  mutate(evolution = "Yes") %>%
  mutate(mpa_size = "10 x 10") %>%
  mutate(climate = "Shocks")
df32 <- read_csv(here::here("generated_data", "summary_data", "summary_m32.csv")) %>%
  mutate(evolution = "No") %>%
  mutate(mpa_size = "10 x 10") %>%
  mutate(climate = "Shocks")

df33 <- read_csv(here::here("generated_data", "summary_data", "summary_m33.csv")) %>%
  mutate(evolution = "Yes") %>%
  mutate(mpa_size = "1 x 1") %>%
  mutate(climate = "Shocks with Mean Shift")
df34 <- read_csv(here::here("generated_data", "summary_data", "summary_m34.csv")) %>%
  mutate(evolution = "No") %>%
  mutate(mpa_size = "1 x 1") %>%
  mutate(climate = "Shocks with Mean Shift")
df35 <- read_csv(here::here("generated_data", "summary_data", "summary_m35.csv")) %>%
  mutate(evolution = "Yes") %>%
  mutate(mpa_size = "3 x 3") %>%
  mutate(climate = "Shocks with Mean Shift")
df36 <- read_csv(here::here("generated_data", "summary_data", "summary_m36.csv")) %>%
  mutate(evolution = "No") %>%
  mutate(mpa_size = "3 x 3") %>%
  mutate(climate = "Shocks with Mean Shift")
df37 <- read_csv(here::here("generated_data", "summary_data", "summary_m37.csv")) %>%
  mutate(evolution = "Yes") %>%
  mutate(mpa_size = "6 x 6") %>%
  mutate(climate = "Shocks with Mean Shift")
df38 <- read_csv(here::here("generated_data", "summary_data", "summary_m38.csv")) %>%
  mutate(evolution = "No") %>%
  mutate(mpa_size = "6 x 6") %>%
  mutate(climate = "Shocks with Mean Shift")
df39 <- read_csv(here::here("generated_data", "summary_data", "summary_m39.csv")) %>%
  mutate(evolution = "Yes") %>%
  mutate(mpa_size = "10 x 10") %>%
  mutate(climate = "Shocks with Mean Shift")
df40 <- read_csv(here::here("generated_data", "summary_data", "summary_m40.csv")) %>%
  mutate(evolution = "No") %>%
  mutate(mpa_size = "10 x 10") %>%
  mutate(climate = "Shocks with Mean Shift")

df41 <- read_csv(here::here("generated_data", "summary_data", "summary_m41.csv")) %>%
  mutate(evolution = "Yes") %>%
  mutate(mpa_size = "None") %>%
  mutate(climate = "Null")
df42 <- read_csv(here::here("generated_data", "summary_data", "summary_m42.csv")) %>%
  mutate(evolution = "No") %>%
  mutate(mpa_size = "None") %>%
  mutate(climate = "Null")
df43 <- read_csv(here::here("generated_data", "summary_data", "summary_m43.csv")) %>%
  mutate(evolution = "Yes") %>%
  mutate(mpa_size = "None") %>%
  mutate(climate = "Mean Shift")
df44 <- read_csv(here::here("generated_data", "summary_data", "summary_m44.csv")) %>%
  mutate(evolution = "No") %>%
  mutate(mpa_size = "None") %>%
  mutate(climate = "Mean Shift")
df45 <- read_csv(here::here("generated_data", "summary_data", "summary_m45.csv")) %>%
  mutate(evolution = "Yes") %>%
  mutate(mpa_size = "None") %>%
  mutate(climate = "El Nino/La Nina")
df46 <- read_csv(here::here("generated_data", "summary_data", "summary_m46.csv")) %>%
  mutate(evolution = "No") %>%
  mutate(mpa_size = "None") %>%
  mutate(climate = "El Nino/La Nina")
df47 <- read_csv(here::here("generated_data", "summary_data", "summary_m47.csv")) %>%
  mutate(evolution = "Yes") %>%
  mutate(mpa_size = "None") %>%
  mutate(climate = "Shocks")
df48 <- read_csv(here::here("generated_data", "summary_data", "summary_m48.csv")) %>%
  mutate(evolution = "No") %>%
  mutate(mpa_size = "None") %>%
  mutate(climate = "Shocks")
df49 <- read_csv(here::here("generated_data", "summary_data", "summary_m49.csv")) %>%
  mutate(evolution = "Yes") %>%
  mutate(mpa_size = "None") %>%
  mutate(climate = "Shocks with Mean Shift")
df50 <- read_csv(here::here("generated_data", "summary_data", "summary_m50.csv")) %>%
  mutate(evolution = "No") %>%
  mutate(mpa_size = "None") %>%
  mutate(climate = "Shocks with Mean Shift")

beepr::beep(5)
```

```{r}
# bind the data together with rbind (can do in groups to conserve memory)
df_list <- list(
  df1, df2, df3, df4, df5, df6, df7, df8, df9, df10,
  df11, df12, df13, df14, df15, df16, df17, df18, df19, df20,
  df21, df22, df23, df24, df25, df26, df27, df28, df29, df30,
  df31, df32, df33, df34, df35, df36, df37, df38, df39, df40,
  df41, df42, df43, df44, df45, df46, df47, df48, df49, df50
) %>%
  reduce(rbind)

# output full summarized data for use in 03_analysis_figs.Rmd
write_csv(df_list, here::here("generated_data", "summary_data", "full_sum.csv"))
```
