==> scripts/04_supplemental.Rmd <==
---
title: "Supplemental Figures and Analysis"
author: "Allie Caughman"
date: "2022-08-11"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(tidyverse)
library(patchwork)
library(plotrix)
```

# Paper Figures 

### Read in Data

```{r}
# read in data
data <- read_csv(here::here("generated_data", "summary_data", "full_sum.csv")) %>%
  mutate(generation = as.numeric(generation)) %>% # make generation a factor
  mutate(climate = fct_relevel(climate, c("Null", "Shocks", "Shocks with Mean Shift", "Mean Shift", "El Nino/La Nina"))) %>% # relevel climate
  mutate(mpa_size = fct_relevel(mpa_size, c("None", "1 x 1", "3 x 3", "6 x 6", "10 x 10"))) # relevel MPA sizes
```

### Data Manipulation

```{r}
# summarizes data within the full simulation
line_df <- data %>%
  filter(generation > 25) %>% # filter generations when MPA is implemented
  group_by(generation, evolution, mpa_size, climate) %>%
  summarise(
    location_sum = sum(pop_mean, na.rm = TRUE), # sum across locations
    max_temp = max(max_temp),
    min_temp = min(min_temp),
    freq_avg = mean(freq, na.rm = TRUE) # average frequency
  ) %>%
  mutate(generation = as.numeric(generation)) # make generation numeric here

# calculates rate of change at each location
change <- data %>%
  select(-genotype, -geno_pop_mean, -freq, -fished_mean, -geno_pop_sd, -geno_pop_se) %>% # remove genotype columns
  distinct() %>%
  group_by(climate, evolution, mpa_size, lat, lon) %>%
  mutate(rate = pop_mean - lag(pop_mean, default = first(pop_mean), order_by = generation)) %>% # calculate fist derivative
  mutate(relative_rate = (pop_mean - lag(pop_mean, default = first(pop_mean), order_by = generation)) / pop_mean) %>% # calculate relative rate of decline
  ungroup()

# summarizes data within the MPA only
line_df_mpa_full <- data %>%
  filter(generation > 25) %>% # filter generations when MPA is implemented
  # filter only MPA grids
  filter((lat %in% c(49:51) & mpa_size == "3 x 3") | (lat %in% c(47:52) & mpa_size == "6 x 6") | (lat %in% c(46:55) & mpa_size == "10 x 10") | (lat %in% c(50) & mpa_size == "1 x 1")) %>%
  filter((lon %in% c(9:11) & mpa_size == "3 x 3") | (lon %in% c(9:14) & mpa_size == "6 x 6") | (lon %in% c(5:14) & mpa_size == "10 x 10") | (lon %in% c(10) & mpa_size == "1 x 1")) %>%
  select(-genotype, -geno_pop_mean, -geno_pop_sd, -geno_pop_se, -freq) %>% # remove genotype columns
  distinct() %>%
  group_by(generation, evolution, mpa_size, climate) %>%
  summarise(
    location_sum = sum(pop_mean, na.rm = TRUE), # sum across location
    max_temp = max(max_temp),
    min_temp = min(min_temp)
  ) %>%
  mutate(generation = as.numeric(generation)) # make generation numeric

by_lat = data %>% 
  filter(generation > 25) %>% # filter generations when MPA is implemented
  select(-genotype, -geno_pop_mean, -geno_pop_sd, -geno_pop_se, -freq) %>% # remove genotype columns
  distinct() %>% 
  mutate(lat_group = case_when(
    lat %in% c(1:10) ~ 1,
    lat %in% c(11:20) ~ 2,
    lat %in% c(21:30) ~ 3,
    lat %in% c(31:40) ~ 4,
    lat %in% c(41:50) ~ 5,
    lat %in% c(51:60) ~ 6,
    lat %in% c(61:70) ~ 7,
    lat %in% c(71:80) ~ 8,
    lat %in% c(81:90) ~ 9,
    lat %in% c(91:100) ~ 10
  )) %>%  # make generation numeric
  mutate(loc = case_when(
    lat_group %in% c(5,6) ~ "MPA",
    !lat_group %in% c(5,6) ~ "non-MPA",
  )) %>% 
  group_by(generation, evolution, mpa_size, climate, lat_group) %>% 
  summarise(
    location_sum = mean(pop_mean, na.rm = TRUE), # sum across location
    max_temp = max(max_temp),
    min_temp = min(min_temp)
  ) %>%
  mutate(generation = as.numeric(generation))
```

```{r}
# calculates mean rate of change within the MPA
mean_change2 <- change %>%
  filter(generation >= 35) %>% # filter generations when climate change causes declines
  # filter only MPA grids
  filter((lat %in% c(49:51) & mpa_size == "3 x 3") | (lat %in% c(47:52) & mpa_size == "6 x 6") | (lat %in% c(46:55) & mpa_size == "10 x 10") | (lat %in% c(50) & mpa_size == "1 x 1")) %>%
  filter((lon %in% c(9:11) & mpa_size == "3 x 3") | (lon %in% c(9:14) & mpa_size == "6 x 6") | (lon %in% c(5:14) & mpa_size == "10 x 10") | (lon %in% c(10) & mpa_size == "1 x 1")) %>%
  filter(!is.infinite(relative_rate)) %>% # remove relative rate where population went to 0
  group_by(generation, climate, evolution, mpa_size) %>%
  summarize(
    mean_change = mean(rate, na.rm = TRUE), # get mean rate of change in MPA
    relative_mean_change = mean(relative_rate, na.rm = TRUE) # get relative rate of change in MPA
  )

collapse_by_lat <- by_lat %>%
  select(generation, evolution, mpa_size, climate, location_sum, lat_group) %>%
  filter(generation > 30) %>% 
  filter(location_sum <= 100 * 2 * 0.01) %>%  
  group_by(evolution, mpa_size, climate, lat_group) %>%
  summarize(min_gen = min(generation))

# calculates collapse points in full simulation
collapse <- line_df %>%
  select(generation, evolution, mpa_size, climate, location_sum) %>%
  filter(generation > 30) %>% # filter generations when climate change causes declines
  filter(location_sum <= 100 * 2 * .05 * 100 * 20) %>% # filter pop to expected collapse level
  group_by(evolution, mpa_size, climate) %>%
  summarize(min_gen = min(generation)) %>% # find the first generation below collapse point
  mutate(collapse_point = 100 * 2 * .05 * 100 * 20) # add collapse point size to column

# calculates collapse points within the MPA only
collapse_full <- line_df_mpa_full %>%
  select(generation, evolution, mpa_size, climate, location_sum) %>%
  filter(generation > 30) %>% # filter generations when climate change causes declines
  filter((location_sum <= (100 * 2 * .05 * 3 * 3) & mpa_size == "3 x 3") | (location_sum <= (100 * 2 * .05 * 6 * 6) & mpa_size == "6 x 6") | (location_sum <= (100 * 2 * .05 * 10 * 10) & mpa_size == "10 x 10") | (location_sum <= (100 * 2 * .05) & mpa_size == "1 x 1")) %>% # calculate collapse rates in different MPAs
  group_by(evolution, mpa_size, climate) %>%
  summarize(min_gen = min(generation)) %>% # find the first generation below collapse point
  mutate(collapse_point = case_when(
    mpa_size == "1 x 1" ~ 100 * 2 * 0.05,
    mpa_size == "3 x 3" ~ 100 * 2 * .05 * 3 * 3,
    mpa_size == "6 x 6" ~ 100 * 2 * .05 * 6 * 6,
    mpa_size == "10 x 10" ~ 100 * 2 * .05 * 10 * 10
  )) # add column with collapse point size
```

## Supplemental Figures

### Main Supplemental Figures

```{r}
# plot popluation through time in full simulation with collapse points
s1 <- ggplot(line_df, aes(generation, location_sum)) +
  geom_vline(data = collapse, aes(xintercept = min_gen, color = climate, linetype = evolution)) +
  geom_hline(data = collapse, aes(yintercept = collapse_point), alpha = 1, color = "darkgrey") +
  geom_line(aes(color = climate, linetype = evolution), size = 0.7) +
  facet_wrap(~mpa_size, scales = "free_y", ncol = 1) +
  theme_bw() +
  labs(
    x = "Time",
    y = "Population Size",
    linetype = "Evolution?",
    color = "Climate"
  ) +
  theme(
    panel.grid.minor = element_blank(),
    strip.text = element_text(face = "bold"),
    strip.background = element_rect(fill = "white"),
    axis.text.x = element_blank(),
    axis.ticks.x = element_blank()
  ) +
  scale_color_manual(values = c("#440154", "#3b528b", "#21918c", "#5ec962", "#fde725")) +
  scale_linetype_manual(values = c("dashed", "solid"))

ggsave(s1, file = paste0("decline_full.pdf"), path = here::here("figs", "supp"), height = 8, width = 15)
```

```{r}
# plot rate of change within the MPA
s3 <- ggplot(mean_change2, aes(generation, mean_change)) +
  geom_line(aes(color = mpa_size, linetype = evolution), linewidth = 0.7) +
  facet_wrap(~climate, ncol = 1, scales = "free_y") +
  theme_bw() +
  labs(
    x = "Time",
    y = "Rate of Population Change",
    linetype = "Evolution?",
    color = "MPA Size"
  ) +
  scale_color_viridis_d() +
  theme(
    panel.grid.minor = element_blank(),
    strip.text = element_text(face = "bold"),
    strip.background = element_rect(fill = "white"),
    axis.text.x = element_blank(),
    axis.ticks.x = element_blank()
  ) +
  scale_linetype_manual(values = c("dashed", "solid"))

ggsave(s3, file = paste0("rate.pdf"), path = here::here("figs", "supp"), height = 10, width = 15)
```

```{r}
# plot population through time within the MPA with collapse points
s4 <- ggplot(line_df_mpa_full, aes(generation, location_sum)) +
  geom_vline(data = collapse_full, aes(xintercept = min_gen, color = climate, linetype = evolution), alpha = 0.8) +
  geom_hline(data = collapse_full, aes(yintercept = collapse_point), alpha = 1, color = "darkgrey") +
  geom_line(aes(color = climate, linetype = evolution), linewidth = 0.7) +
  facet_wrap(~mpa_size, scales = "free", ncol = 1) +
  theme_bw() +
  labs(
    x = "Time",
    y = "Population Size",
    color = "Climate", fill = "", linetype = "Evolution?"
  ) +
  theme(
    panel.grid.minor = element_blank(),
    strip.text = element_text(face = "bold"),
    strip.background = element_rect(fill = "white"),
    axis.text.x = element_blank(),
    axis.ticks.x = element_blank()
  ) +
  scale_color_viridis_d() +
  scale_fill_viridis_d() +
  guides(fill = "none") +
  scale_linetype_manual(values = c("dashed", "solid"))

ggsave(s4, file = paste0("decline_mpa_full.pdf"), path = here::here("figs", "supp"), height = 8, width = 15)
```

```{r}
# plot the natural log of population within the full simulation
s5 <- ggplot(line_df, aes(generation, log(location_sum))) +
  geom_line(aes(color = climate, linetype = evolution), linewidth = 0.7) +
  facet_wrap(~mpa_size, ncol = 1) +
  theme_bw() +
  labs(
    x = "Time",
    y = "Population Size",
    color = "Climate",
    linetype = "Evolution?"
  ) +
  theme(
    panel.grid.minor = element_blank(),
    strip.text = element_text(face = "bold"),
    strip.background = element_rect(fill = "white"),
    axis.text.x = element_blank(),
    axis.ticks.x = element_blank()
  ) +
  scale_color_manual(values = c("#21918c", "#5ec962", "#fde725", "#3b528b", "#440154")) +
  scale_linetype_manual(values = c("dashed", "solid"))

ggsave(s5, file = paste0("decline_full_sim_log.pdf"), path = here::here("figs", "supp"), height = 8, width = 15)
```


```{r}
med2 = by_lat %>% 
  filter(mpa_size == "6 x 6")

# plot population through time in the Medium MPA
s6 <- ggplot(med2, aes(generation, location_sum)) +
  geom_line(aes(linetype = evolution, color = evolution), linewidth = 0.7) +
  theme_bw(base_size = 12) +
  ggh4x::facet_grid2(desc(lat_group)~climate, scales = "free_y", independent = "y") +
  labs(
    x = "Time",
    y = "Population Size",
    color = "Climate Scenario",
    linetype = "Evolution?"
  ) +
  theme(
    panel.grid.minor = element_blank(),
    strip.text = element_text(face = "bold"),
    strip.background = element_rect(fill = "white"),
    axis.text.x = element_blank(),
    axis.ticks.x = element_blank()
  ) +
  scale_color_manual(values = c("#21918c", "#440154")) +
  scale_linetype_manual(values = c("dashed", "solid"))

ggsave(s6, file = paste0("pop_by_lat.pdf"), path = here::here("figs", "sup"), height = 15, width = 25)
```

```{r}
# plot the collapse points of the population
s7 <- ggplot(collapse_by_lat, aes(min_gen, mpa_size)) +
  geom_point(aes(color = climate, shape = evolution), size = 3) +
  theme_bw(base_size = 12) +
  scale_color_manual(values = c("#21918c", "#5ec962", "#fde725", "#3b528b", "#440154")) +
  labs(
    x = "Time",
    y = "MPA Size",
    color = "Climate Scenario",
    shape = "Evolution?"
  ) +
  theme(
    panel.grid.minor = element_blank(),
    axis.text.x = element_blank(),
    axis.ticks.x = element_blank(),
    strip.text = element_text(face = "bold"),
    strip.background = element_rect(fill = "white")
  ) +
  facet_wrap(~desc(lat_group), ncol = 1)

ggsave(s7, file = paste0("by_lat_collapse.pdf"), path = here::here("figs", "supp"), height = 25, width = 20)
```



### Climate Rate Figures

```{r}
# read in data that compares different climate rates (data not in github, but simulations can be run and merged similarly to regular data)

data_rate <- read_csv(here::here("generated_data", "summary_data", "full_sum_climate_rate.csv")) %>%
  mutate(generation = as.numeric(generation)) %>% # generation as numeric
  mutate(climate = fct_relevel(climate, c("Shocks with Mean Shift", "Mean Shift", "El Nino/La Nina"))) %>% # relevel climate
  mutate(mpa_size = fct_relevel(mpa_size, c("3 x 3", "6 x 6", "10 x 10"))) # relevel MPA sizes
```

```{r}
# functions preform the same functions seen in 03_analysis_figs.Rmd for analyses the main text data

line_df_mpa_full_rate <- data_rate %>%
  filter(generation > 25) %>%
  filter((lat %in% c(49:51) & mpa_size == "3 x 3") | (lat %in% c(47:52) & mpa_size == "6 x 6") | (lat %in% c(46:55) & mpa_size == "10 x 10")) %>%
  filter((lon %in% c(9:11) & mpa_size == "3 x 3") | (lon %in% c(9:14) & mpa_size == "6 x 6") | (lon %in% c(5:14) & mpa_size == "10 x 10")) %>%
  select(-genotype, -geno_pop_mean, -geno_pop_sd, -geno_pop_se, -freq) %>%
  distinct() %>%
  group_by(generation, evolution, mpa_size, climate, climate_rate) %>%
  summarise(
    location_sum = sum(pop_mean, na.rm = TRUE),
  ) %>%
  mutate(generation = as.numeric(generation))

collapse_mpa_rate <- line_df_mpa_full_rate %>%
  select(generation, evolution, mpa_size, climate, location_sum, climate_rate) %>%
  filter(generation > 30) %>%
  filter((location_sum <= (100 * 2 * .05 * 3 * 3) & mpa_size == "3 x 3") | (location_sum <= (100 * 2 * .05 * 6 * 6) & mpa_size == "6 x 6") | (location_sum <= (100 * 2 * .05 * 10 * 10) & mpa_size == "10 x 10")) %>%
  group_by(evolution, mpa_size, climate, climate_rate) %>%
  summarize(min_gen = min(generation)) %>%
  mutate(collapse_point = case_when(
    mpa_size == "3 x 3" ~ 100 * 2 * .05 * 3 * 3,
    mpa_size == "6 x 6" ~ 100 * 2 * .05 * 6 * 6,
    mpa_size == "10 x 10" ~ 100 * 2 * .05 * 10 * 10
  )) %>%
  mutate(area = "mpa only")

write_csv(collapse_mpa_rate, here::here("generated_data", "results", "collapse_climate_rate.csv"))

change_rate <- data_rate %>%
  select(-genotype, -geno_pop_mean, -freq, -fished_mean, -geno_pop_sd, -geno_pop_se) %>%
  distinct() %>%
  group_by(climate, evolution, mpa_size, climate_rate, lat, lon) %>%
  mutate(rate = pop_mean - lag(pop_mean, default = first(pop_mean), order_by = generation)) %>%
  ungroup()

change_mpa_rate <- change_rate %>%
  filter(generation >= 50) %>%
  filter((lat %in% c(49:51) & mpa_size == "3 x 3") | (lat %in% c(47:52) & mpa_size == "6 x 6") | (lat %in% c(46:55) & mpa_size == "10 x 10")) %>%
  filter((lon %in% c(9:11) & mpa_size == "3 x 3") | (lon %in% c(9:14) & mpa_size == "6 x 6") | (lon %in% c(5:14) & mpa_size == "10 x 10")) %>%
  group_by(climate, evolution, mpa_size, climate_rate) %>%
  summarize(mean_change = mean(rate, na.rm = TRUE))

write_csv(change_mpa_rate, here::here("generated_data", "results", "climate_rate_change_mpa.csv"))

line_df_geno_rate <- data_rate %>%
  filter((lat %in% c(49:51) & mpa_size == "3 x 3") | (lat %in% c(47:52) & mpa_size == "6 x 6") | (lat %in% c(46:55) & mpa_size == "10 x 10") | (lat %in% c(50) & mpa_size == "1 x 1")) %>%
  filter((lon %in% c(9:11) & mpa_size == "3 x 3") | (lon %in% c(9:14) & mpa_size == "6 x 6") | (lon %in% c(5:14) & mpa_size == "10 x 10") | (lon %in% c(10) & mpa_size == "1 x 1")) %>%
  group_by(generation, evolution, mpa_size, climate, genotype, climate_rate) %>%
  summarise(
    location_sum = sum(geno_pop_mean, na.rm = TRUE),
    max_temp = max(max_temp),
    min_temp = min(min_temp),
    freq_avg = mean(freq, na.rm = TRUE)
  ) %>%
  mutate(generation = as.numeric(generation)) %>%
  mutate(min_m = 1 - (1 - exp((-(max_temp - 25)^2) / (4^2)))) %>%
  mutate(min_survival = ifelse(min_m > 0.7, 0.7, min_m)) %>%
  mutate(max_m = 1 - (1 - exp((-(min_temp - 25)^2) / (4^2)))) %>%
  mutate(max_survival = ifelse(max_m > 0.7, 0.7, max_m)) %>%
  select(-max_m, -min_m) %>%
  filter(evolution == "Yes")

aa_geno_rate <- line_df_geno_rate %>%
  filter(genotype == "aa") %>%
  filter(climate %in% c("Mean Shift", "Shocks with Mean Shift", "El Nino/La Nina")) %>%
  filter(!is.na(freq_avg)) %>%
  ungroup() %>%
  select(-evolution, -genotype) %>%
  mutate(
    average_survival = max_survival + min_survival,
    average_temp = max_temp + min_temp
  ) %>%
  mutate(
    average_survival = average_survival / 2,
    average_temp = average_temp / 2
  )

aa_sum_rate <- aa_geno_rate %>%
  group_by(mpa_size, climate, climate_rate) %>%
  summarize(freq_avg = max(freq_avg))

aa_rate <- left_join(aa_sum_rate, aa_geno_rate)

temps_rate <- aa_rate %>%
  select(mpa_size, climate, freq_avg, average_temp, generation, average_survival, climate_rate)

write_csv(temps_rate, here::here("generated_data", "results", "aa_peak_climate_rate.csv"))

line_df_geno_lat_rate <- data_rate %>%
  group_by(generation, evolution, mpa_size, climate, genotype, lat) %>%
  summarise(
    location_sum = sum(geno_pop_mean, na.rm = TRUE),
    freq_avg = mean(freq, na.rm = TRUE)
  ) %>%
  mutate(generation = as.numeric(generation)) %>%
  filter(evolution == "Yes") %>%
  mutate(lat = as.factor(lat)) 

aa_geno2_rate <- line_df_geno_lat_rate %>%
  ungroup() %>%
  filter(genotype == "aa") %>%
  filter(lat == 50) %>%
  filter(climate %in% c("Mean Shift", "Shocks with Mean Shift", "El Nino/La Nina")) %>%
  filter(!is.na(freq_avg)) %>%
  select(-evolution, -genotype)

aa_mean_rate <- aa_geno2_rate %>%
  filter(climate == "Mean Shift") %>%
  filter(freq_avg < 0.9)
aa_mean_shock_rate <- aa_geno2_rate %>%
  filter(climate == "Shocks with Mean Shift") %>%
  filter(freq_avg < 0.9)
aa_enso_rate <- aa_geno2_rate %>%
  filter(climate == "El Nino/La Nina") %>%
  filter(freq_avg < 0.9)

line_df_mpa_full_rate <- data_rate %>%
  filter(generation > 25) %>%
  filter((lat %in% c(49:51) & mpa_size == "3 x 3") | (lat %in% c(47:52) & mpa_size == "6 x 6") | (lat %in% c(46:55) & mpa_size == "10 x 10") | (lat %in% c(50) & mpa_size == "1 x 1")) %>%
  filter((lon %in% c(9:11) & mpa_size == "3 x 3") | (lon %in% c(9:14) & mpa_size == "6 x 6") | (lon %in% c(10) & mpa_size == "10 x 10")) %>%
  select(-genotype, -geno_pop_mean, -geno_pop_sd, -geno_pop_se, -freq) %>%
  distinct()

line_df_geno_lat_rate <- data_rate %>%
  filter(lat == 10 | lat == 90 | lat == 50)

line_df_mpa_full_rate <- line_df_mpa_full_rate %>%
  group_by(generation, evolution, mpa_size, climate, climate_rate) %>%
  summarise(
    location_sum = sum(pop_mean, na.rm = TRUE)
  ) %>%
  mutate(generation = as.numeric(generation))

line_df_geno_lat_rate <- line_df_geno_lat_rate %>%
  group_by(generation, evolution, mpa_size, climate, genotype, lat, climate_rate) %>%
  summarise(
    location_sum = sum(geno_pop_mean, na.rm = TRUE),
    freq_avg = mean(freq, na.rm = TRUE)
  ) %>%
  mutate(generation = as.numeric(generation)) %>%
  filter(evolution == "Yes") %>%
  mutate(lat = as.factor(lat))

line_lat_mpa_combo_rate <- line_df_geno_lat_rate %>%
  filter(lat != 50) %>%
  group_by(generation, climate, genotype, lat, climate_rate) %>%
  summarize(mpa_mean_freq = mean(freq_avg, na.rm = TRUE)) %>%
  mutate(location = case_when(
    lat == 10 ~ "Low Latitude",
    lat == 90 ~ "High Latitude"
  )) %>%
  mutate(mpa_size = "Non-MPA")

geno_lat_rate <- line_df_geno_lat_rate %>%
  filter(lat == 50) %>%
  mutate(location = "MPA") %>%
  mutate(mpa_mean_freq = freq_avg) %>%
  full_join(line_lat_mpa_combo_rate) %>%
  mutate(location = as.factor(location)) %>%
  mutate(mpa_size = as.factor(mpa_size)) %>%
  mutate(mpa_size = fct_relevel(mpa_size, c("10 x 10", "6 x 6", "3 x 3", "Non-MPA"))) %>%
  filter(mpa_size != "Non-MPA")

collapse_full_rate <- line_df_mpa_full_rate %>%
  select(generation, evolution, mpa_size, climate, location_sum, climate_rate) %>%
  filter(generation > 30) %>%
  filter((location_sum <= (100 * 2 * .05 * 3 * 3) & mpa_size == "3 x 3") | (location_sum <= (100 * 2 * .05 * 6 * 6) & mpa_size == "6 x 6") | (location_sum <= (100 * 2 * .05 * 10 * 10) & mpa_size == "10 x 10") | (location_sum <= (100 * 2 * .05) & mpa_size == "1 x 1")) %>%
  group_by(evolution, mpa_size, climate, climate_rate) %>%
  summarize(min_gen = min(generation)) %>%
  mutate(collapse_point = case_when(
    mpa_size == "1 x 1" ~ 100 * 2 * 0.05,
    mpa_size == "3 x 3" ~ 100 * 2 * .05 * 3 * 3,
    mpa_size == "6 x 6" ~ 100 * 2 * .05 * 6 * 6,
    mpa_size == "10 x 10" ~ 100 * 2 * .05 * 10 * 10
  ))
```

```{r}
# plot population size through time with collapse points
s6 <- ggplot(line_df_mpa_full_rate, aes(generation, location_sum)) +
  geom_vline(data = collapse_full_rate, aes(xintercept = min_gen, color = as.factor(climate_rate), linetype = evolution), alpha = 0.8) +
  geom_hline(data = collapse_full_rate, aes(yintercept = collapse_point), alpha = 1, color = "darkgrey") +
  geom_line(aes(color = as.factor(climate_rate), linetype = evolution)) +
  ggh4x::facet_grid2(mpa_size ~ climate, scales = "free_y", independent = "y") +
  theme_bw() +
  labs(
    x = "Time",
    y = "Population Size",
    color = "Climate Increase Per Year", linetype = "Evolution?"
  ) +
  theme(
    panel.grid.minor = element_blank(),
    strip.text = element_text(face = "bold"),
    strip.background = element_rect(fill = "white"),
    axis.text.x = element_blank(),
    axis.ticks.x = element_blank()
  ) +
  scale_color_manual(values = c("#5ec962", "#440154")) +
  guides(fill = "none")

# ggsave(s6, file = paste0("decline_mpa_climate_rate.pdf"), path = here::here("figs", "supp"), height = 8, width = 15)
```

```{r}
# plot genotype frequencies through time
s7 <- ggplot(geno_lat_rate, aes(generation, mpa_mean_freq)) +
  geom_line(aes(color = as.factor(climate_rate), linetype = mpa_size, alpha = as.factor(climate_rate))) +
  facet_grid(genotype ~ climate, scales = "free_y") +
  theme_bw() +
  labs(
    x = "Time",
    y = "Genotype Frequency",
    color = "Climate Increase Per Year",
    linetype = "MPA size",
    alpha = ""
  ) +
  theme() +
  scale_color_manual(values = c("#5ec962", "#440154")) +
  theme(
    panel.grid.minor = element_blank(),
    strip.text = element_text(face = "bold"),
    strip.background = element_rect(fill = "white"),
    axis.text.x = element_blank(),
    axis.ticks.x = element_blank()
  ) +
  scale_linetype_manual(values = c("solid", "dashed", "dotted")) +
  scale_alpha_manual(values = c(1, 0.3), guide = "none")

s8 <- s6 + s7 + plot_annotation(tag_levels = "A") + plot_layout(guides = "collect")

ggsave(s8, file = paste0("climate_rate.pdf"), path = here::here("figs", "supp"), height = 8, width = 15)

# ggsave(s7, file = paste0("genotype_summary_rate.pdf"), path = here::here("figs", "supp"), height = 8, width = 15)
```

### Fishing Pressure Figures

```{r}
# read in data that compares different fishing pressures (data not in github, but simulations can be run and merged similarly to regular data)

data_fp <- read_csv(here::here("generated_data", "summary_data", "full_sum_fishing_pressure.csv")) %>%
  mutate(generation = as.numeric(generation)) %>%
  mutate(climate = fct_relevel(climate, c("Shocks with Mean Shift", "Mean Shift", "El Nino/La Nina"))) %>%
  filter(generation <= 150) %>%
  mutate(mpa_size = fct_relevel(mpa_size, c("3 x 3", "6 x 6", "10 x 10"))) # relevel MPA sizes
```

```{r}
# functions preform the same functions seen in 03_analysis_figs.Rmd for analyses the main text data

line_df_mpa_full_fp <- data_fp %>%
  filter(generation > 25) %>%
  filter((lat %in% c(49:51) & mpa_size == "3 x 3") | (lat %in% c(47:52) & mpa_size == "6 x 6") | (lat %in% c(46:55) & mpa_size == "10 x 10")) %>%
  filter((lon %in% c(9:11) & mpa_size == "3 x 3") | (lon %in% c(9:14) & mpa_size == "6 x 6") | (lon %in% c(5:14) & mpa_size == "10 x 10")) %>%
  select(-genotype, -geno_pop_mean, -geno_pop_sd, -geno_pop_se, -freq) %>%
  distinct() %>%
  group_by(generation, evolution, mpa_size, climate, fishing_pressure) %>%
  summarise(
    location_sum = sum(pop_mean, na.rm = TRUE)
  ) %>%
  mutate(generation = as.numeric(generation))

collapse_mpa_fp <- line_df_mpa_full_fp %>%
  select(generation, evolution, mpa_size, climate, location_sum, fishing_pressure) %>%
  filter(generation > 30) %>%
  filter((location_sum <= (100 * 2 * .05 * 3 * 3) & mpa_size == "3 x 3") | (location_sum <= (100 * 2 * .05 * 6 * 6) & mpa_size == "6 x 6") | (location_sum <= (100 * 2 * .05 * 10 * 10) & mpa_size == "10 x 10")) %>%
  group_by(evolution, mpa_size, climate, fishing_pressure) %>%
  summarize(min_gen = min(generation)) %>%
  mutate(collapse_point = case_when(
    mpa_size == "3 x 3" ~ 100 * 2 * .05 * 3 * 3,
    mpa_size == "6 x 6" ~ 100 * 2 * .05 * 6 * 6,
    mpa_size == "10 x 10" ~ 100 * 2 * .05 * 10 * 10
  )) %>%
  mutate(area = "mpa only")

write_csv(collapse_mpa_fp, here::here("generated_data", "results", "collapse_fishing_pressure.csv"))

change_fp <- data_fp %>%
  select(-genotype, -geno_pop_mean, -freq, -fished_mean, -geno_pop_sd, -geno_pop_se) %>%
  distinct() %>%
  group_by(climate, evolution, mpa_size, fishing_pressure, lat, lon) %>%
  mutate(rate = pop_mean - lag(pop_mean, default = first(pop_mean), order_by = generation)) %>%
  ungroup()

change_mpa_fp <- change_fp %>%
  filter(generation >= 50) %>%
  filter((lat %in% c(49:51) & mpa_size == "3 x 3") | (lat %in% c(47:52) & mpa_size == "6 x 6") | (lat %in% c(46:55) & mpa_size == "10 x 10")) %>%
  filter((lon %in% c(9:11) & mpa_size == "3 x 3") | (lon %in% c(9:14) & mpa_size == "6 x 6") | (lon %in% c(5:14) & mpa_size == "10 x 10")) %>%
  group_by(climate, evolution, mpa_size, fishing_pressure) %>%
  summarize(mean_change = mean(rate, na.rm = TRUE))

write_csv(change_mpa_fp, here::here("generated_data", "results", "fishing_pressure_change_mpa.csv"))

line_df_geno_fp <- data_fp %>%
  filter((lat %in% c(49:51) & mpa_size == "3 x 3") | (lat %in% c(47:52) & mpa_size == "6 x 6") | (lat %in% c(46:55) & mpa_size == "10 x 10") | (lat %in% c(50) & mpa_size == "1 x 1")) %>%
  filter((lon %in% c(9:11) & mpa_size == "3 x 3") | (lon %in% c(9:14) & mpa_size == "6 x 6") | (lon %in% c(5:14) & mpa_size == "10 x 10") | (lon %in% c(10) & mpa_size == "1 x 1")) %>%
  group_by(generation, evolution, mpa_size, climate, genotype, fishing_pressure) %>%
  summarise(
    location_sum = sum(geno_pop_mean, na.rm = TRUE),
    freq_avg = mean(freq, na.rm = TRUE)
  ) %>%
  mutate(generation = as.numeric(generation)) %>%
  filter(evolution == "Yes")

aa_geno_fp <- line_df_geno_fp %>%
  filter(genotype == "aa") %>%
  filter(climate %in% c("Mean Shift", "Shocks with Mean Shift", "El Nino/La Nina")) %>%
  filter(!is.na(freq_avg)) %>%
  ungroup() %>% 
  select(-evolution, -genotype)

aa_sum_fp <- aa_geno_fp %>%
  group_by(mpa_size, climate, fishing_pressure) %>%
  summarize(freq_avg = max(freq_avg))

aa_fp <- left_join(aa_sum_fp, aa_geno_fp)

temps_fp <- aa_fp %>%
  select(mpa_size, climate, freq_avg, generation, fishing_pressure)

write_csv(temps_fp, here::here("generated_data", "results", "aa_peak_fishing_pressure.csv"))

line_df_fp <- data_fp %>%
  filter(generation > 25) %>%
  group_by(generation, mpa_size, climate, fishing_pressure, evolution) %>%
  summarise(
    location_sum = sum(pop_mean, na.rm = TRUE),
    freq_avg = mean(freq, na.rm = TRUE)
  ) %>%
  mutate(generation = as.numeric(generation))

line_df_geno_lat_fp <- data_fp %>%
  group_by(generation, evolution, mpa_size, climate, genotype, lat, fishing_pressure) %>%
  summarise(
    location_sum = sum(geno_pop_mean, na.rm = TRUE),
    freq_avg = mean(freq, na.rm = TRUE)
  ) %>%
  mutate(generation = as.numeric(generation)) %>%
  filter(evolution == "Yes") %>%
  mutate(lat = as.factor(lat)) 

line_df_geno_lat_fp <- line_df_geno_lat_fp %>%
  filter(lat == 10 | lat == 90 | lat == 50)

line_lat_mpa_combo_fp <- line_df_geno_lat_fp %>%
  filter(lat != 50) %>%
  group_by(generation, climate, genotype, lat, fishing_pressure, evolution) %>%
  summarize(mpa_mean_freq = mean(freq_avg, na.rm = TRUE)) %>%
  mutate(location = case_when(
    lat == 10 ~ "Low Latitude",
    lat == 90 ~ "High Latitude"
  )) %>%
  mutate(mpa_size = "Non-MPA")

geno_lat_fp <- line_df_geno_lat_fp %>%
  filter(lat == 50) %>%
  mutate(location = "MPA") %>%
  mutate(mpa_mean_freq = freq_avg) %>%
  full_join(line_lat_mpa_combo_fp) %>%
  mutate(location = as.factor(location)) %>%
  mutate(mpa_size = as.factor(mpa_size)) %>%
  mutate(mpa_size = fct_relevel(mpa_size, c("10 x 10", "6 x 6", "3 x 3", "Non-MPA"))) %>%
  filter(mpa_size != "Non-MPA")

collapse_fp <- line_df_fp %>%
  select(generation, mpa_size, climate, location_sum, fishing_pressure, evolution) %>%
  filter(generation > 30) %>%
  filter(location_sum <= 100 * 2 * .05 * 100 * 20) %>%
  group_by(mpa_size, climate, fishing_pressure, evolution) %>%
  summarize(min_gen = min(generation)) %>%
  mutate(collapse_point = 100 * 2 * .05 * 100 * 20)

collapse_full_fp <- line_df_mpa_full_fp %>%
  select(generation, evolution, mpa_size, climate, location_sum, fishing_pressure) %>%
  filter(generation > 30) %>%
  filter((location_sum <= (100 * 2 * .05 * 3 * 3) & mpa_size == "3 x 3") | (location_sum <= (100 * 2 * .05 * 6 * 6) & mpa_size == "6 x 6") | (location_sum <= (100 * 2 * .05 * 10 * 10) & mpa_size == "10 x 10") | (location_sum <= (100 * 2 * .05) & mpa_size == "1 x 1")) %>%
  group_by(evolution, mpa_size, climate, fishing_pressure) %>%
  summarize(min_gen = min(generation)) %>%
  mutate(collapse_point = case_when(
    mpa_size == "1 x 1" ~ 100 * 2 * .05,
    mpa_size == "3 x 3" ~ 100 * 2 * .05 * 3 * 3,
    mpa_size == "6 x 6" ~ 100 * 2 * .05 * 6 * 6,
    mpa_size == "10 x 10" ~ 100 * 2 * .05 * 10 * 10
  ))
```


```{r}
# plot genotype frequencies through time
s9 <- ggplot(geno_lat_fp, aes(generation, mpa_mean_freq)) +
  geom_line(aes(color = as.factor(fishing_pressure), linetype = mpa_size, alpha = as.factor(fishing_pressure))) +
  facet_grid(genotype ~ climate, scales = "free_y") +
  theme_bw() +
  labs(
    x = "Time",
    y = "Genotype Frequency",
    color = "Fishing Pressure",
    linetype = "MPA size",
    alpha = ""
  ) +
  theme() +
  scale_color_manual(values = c("#5ec962", "#440154")) +
  theme(
    panel.grid.minor = element_blank(),
    strip.text = element_text(face = "bold"),
    strip.background = element_rect(fill = "white"),
    axis.text.x = element_blank(),
    axis.ticks.x = element_blank()
  ) +
  scale_linetype_manual(values = c("solid", "dashed", "dotted")) +
  scale_alpha_manual(values = c(1, 0.3), guide = "none")

# ggsave(s9, file = paste0("genotype_summary_fishing_pressure.pdf"), path = here::here("figs", "supp"), height = 8, width = 15)
```

```{r}
# plot population size through time with collapse points
s10 <- ggplot(line_df_mpa_full_fp, aes(generation, location_sum)) +
  geom_vline(data = collapse_full_fp, aes(xintercept = min_gen, color = as.factor(fishing_pressure), linetype = evolution), alpha = 0.8) +
  geom_hline(data = collapse_full_fp, aes(yintercept = collapse_point), alpha = 1, color = "darkgrey") +
  geom_line(aes(color = as.factor(fishing_pressure), linetype = evolution)) +
  ggh4x::facet_grid2(mpa_size ~ climate, scales = "free_y", independent = "y") +
  theme_bw() +
  labs(
    x = "Time",
    y = "Population Size",
    color = "Fishing Pressure", linetype = "Evolution?"
  ) +
  theme(
    panel.grid.minor = element_blank(),
    strip.text = element_text(face = "bold"),
    strip.background = element_rect(fill = "white"),
    axis.text.x = element_blank(),
    axis.ticks.x = element_blank()
  ) +
  scale_color_manual(values = c("#5ec962", "#440154"))

s11 <- s10 + s9 + plot_annotation(tag_levels = "A") + plot_layout(guides = "collect")

ggsave(s11, file = paste0("fishing_pressure.pdf"), path = here::here("figs", "supp"), height = 8, width = 15)

# ggsave(s10, file = paste0("decline_mpa_fishing_pressure.pdf"), path = here::here("figs", "supp"), height = 8, width = 15)
```

## Climate Figures

```{r}
years <- 175
NS.patches <- 100
EW.patches <- 20
opt.temp <- 25

# same function from 01_final_model.R
calc_temp_mortality <- function(SST, opt.temp, temp.range, s) {
  nat.m <- array(0, c(nrow(SST), ncol(SST)))
  m <- 1 - exp((-(SST[, ] - opt.temp)^2) / (temp.range^2)) # temperature based mortality function from Walsworth et al.
  m <- 1 - m
  nat.m <- ifelse(m > s, s, m)
  return(nat.m)
}

# set up the same as in 01_final_model.R
SST.patches.mean <- array(opt.temp, c(NS.patches, EW.patches, years))
start_SST <- (opt.temp) + NS.patches * 0.02

for (i in 1:25) {
  SST <- start_SST
  for (lat in 1:NS.patches) {
    SST.patches.mean[lat, , i] <- SST
    SST <- SST - 0.02
  }
}

for (i in 26:years) {
  SST <- start_SST
  for (lat in 1:NS.patches) {
    SST.patches.mean[lat, , i] <- SST
    SST <- SST - 0.02
  }
  start_SST <- start_SST + 0.033
}

SST.patches.enso <- array(opt.temp, c(NS.patches, EW.patches, years))
start_SST <- (opt.temp) + NS.patches * 0.02

for (i in 1:25) {
  SST <- start_SST
  for (lat in 1:NS.patches) {
    SST.patches.enso[lat, , i] <- SST
    SST <- SST - 0.02
  }
}

t <- seq(1, years - 25, 1)
b <- array(0, 25)
change <- 0.5 * sin(t) + 0.033
enso.value <- c(b, change)

for (i in 26:years) {
  SST <- start_SST
  for (lat in 1:NS.patches) {
    SST.patches.enso[lat, , i] <- SST
    SST <- SST - 0.02
  }
  start_SST <- start_SST + enso.value[i]
}

SST.patches.shock <- array(opt.temp, c(NS.patches, EW.patches, years))
start_SST <- (opt.temp) + NS.patches * 0.02

for (i in 1:25) {
  SST <- start_SST
  for (lat in 1:NS.patches) {
    SST.patches.shock[lat, , i] <- SST
    SST <- SST - 0.02
  }
}

for (i in 26:years) {
  SST <- start_SST
  heat_prob <- runif(1, 0, 1)
  if ((i < 75 & heat_prob < 0.1) | (i >= 75 & heat_prob < 0.35)) {
    intensity <- runif(1, 1, 4)
    SST <- start_SST + intensity
  } else {
    SST <- start_SST
  }
  for (lat in 1:NS.patches) {
    SST.patches.shock[lat, , i] <- SST
    SST <- SST - 0.02
  }
}

SST.patches.mean.shock <- array(opt.temp, c(NS.patches, EW.patches, years))
start_SST <- (opt.temp) + NS.patches * 0.02

for (i in 1:25) {
  SST <- start_SST
  for (lat in 1:NS.patches) {
    SST.patches.mean.shock[lat, , i] <- SST
    SST <- SST - 0.02
  }
}

for (i in 26:years) {
  heat_prob <- runif(1, 0, 1)
  if ((i < 75 & heat_prob < 0.1) | (i >= 75 & heat_prob < 0.35)) {
    intensity <- runif(1, 1, 4)
    SST <- start_SST + intensity
  } else {
    SST <- start_SST
  }
  for (lat in 1:NS.patches) {
    SST.patches.mean.shock[lat, , i] <- SST
    SST <- SST - 0.02
  }
  start_SST <- start_SST + 0.033
}

# create dataframes to hold results
output_df_mean <- data.frame()
output_df_enso <- data.frame()
output_df_shock <- data.frame()
output_df_mean_shock <- data.frame()


### Convert into dataframes
# Mean --------------------------------------------------------------------

for (b in 1:years) {
  SSTdf_mean <- SST.patches.mean[, , b] %>%
    as.data.frame()

  SSTdf_mean$lat <- c(1:NS.patches)
  SSTdf_mean$year <- paste0(b)

  output_df_mean <- bind_rows(output_df_mean, SSTdf_mean)
}

SSTdf_mean <- output_df_mean %>%
  pivot_longer(V1:V20,
    names_to = "lon",
    values_to = "sst"
  ) %>%
  mutate(lon = case_when(
    lon == "V1" ~ 1,
    lon == "V2" ~ 2,
    lon == "V3" ~ 3,
    lon == "V4" ~ 4,
    lon == "V5" ~ 5,
    lon == "V6" ~ 6,
    lon == "V7" ~ 7,
    lon == "V8" ~ 8,
    lon == "V9" ~ 9,
    lon == "V10" ~ 10,
    lon == "V11" ~ 11,
    lon == "V12" ~ 12,
    lon == "V13" ~ 13,
    lon == "V14" ~ 14,
    lon == "V15" ~ 15,
    lon == "V16" ~ 16,
    lon == "V17" ~ 17,
    lon == "V18" ~ 18,
    lon == "V19" ~ 19,
    lon == "V20" ~ 20
  )) %>%
  mutate(year = as.numeric(year))

for (b in 1:years) {
  SSTdf_enso <- SST.patches.enso[, , b] %>%
    as.data.frame()

  SSTdf_enso$lat <- c(1:NS.patches)
  SSTdf_enso$year <- paste0(b)

  output_df_enso <- bind_rows(output_df_enso, SSTdf_enso)
}

SSTdf_enso <- output_df_enso %>%
  pivot_longer(V1:V20,
    names_to = "lon",
    values_to = "sst"
  ) %>%
  mutate(lon = case_when(
    lon == "V1" ~ 1,
    lon == "V2" ~ 2,
    lon == "V3" ~ 3,
    lon == "V4" ~ 4,
    lon == "V5" ~ 5,
    lon == "V6" ~ 6,
    lon == "V7" ~ 7,
    lon == "V8" ~ 8,
    lon == "V9" ~ 9,
    lon == "V10" ~ 10,
    lon == "V11" ~ 11,
    lon == "V12" ~ 12,
    lon == "V13" ~ 13,
    lon == "V14" ~ 14,
    lon == "V15" ~ 15,
    lon == "V16" ~ 16,
    lon == "V17" ~ 17,
    lon == "V18" ~ 18,
    lon == "V19" ~ 19,
    lon == "V20" ~ 20
  )) %>%
  mutate(year = as.numeric(year))

for (b in 1:years) {
  SSTdf_shock <- SST.patches.shock[, , b] %>%
    as.data.frame()

  SSTdf_shock$lat <- c(1:NS.patches)
  SSTdf_shock$year <- paste0(b)

  output_df_shock <- bind_rows(output_df_shock, SSTdf_shock)
}

SSTdf_shock <- output_df_shock %>%
  pivot_longer(V1:V20,
    names_to = "lon",
    values_to = "sst"
  ) %>%
  mutate(lon = case_when(
    lon == "V1" ~ 1,
    lon == "V2" ~ 2,
    lon == "V3" ~ 3,
    lon == "V4" ~ 4,
    lon == "V5" ~ 5,
    lon == "V6" ~ 6,
    lon == "V7" ~ 7,
    lon == "V8" ~ 8,
    lon == "V9" ~ 9,
    lon == "V10" ~ 10,
    lon == "V11" ~ 11,
    lon == "V12" ~ 12,
    lon == "V13" ~ 13,
    lon == "V14" ~ 14,
    lon == "V15" ~ 15,
    lon == "V16" ~ 16,
    lon == "V17" ~ 17,
    lon == "V18" ~ 18,
    lon == "V19" ~ 19,
    lon == "V20" ~ 20
  )) %>%
  mutate(year = as.numeric(year))

for (b in 1:years) {
  SSTdf_mean_shock <- SST.patches.mean.shock[, , b] %>%
    as.data.frame()

  SSTdf_mean_shock$lat <- c(1:NS.patches)
  SSTdf_mean_shock$year <- paste0(b)

  output_df_mean_shock <- bind_rows(output_df_mean_shock, SSTdf_mean_shock)
}

SSTdf_mean_shock <- output_df_mean_shock %>%
  pivot_longer(V1:V20,
    names_to = "lon",
    values_to = "sst"
  ) %>%
  mutate(lon = case_when(
    lon == "V1" ~ 1,
    lon == "V2" ~ 2,
    lon == "V3" ~ 3,
    lon == "V4" ~ 4,
    lon == "V5" ~ 5,
    lon == "V6" ~ 6,
    lon == "V7" ~ 7,
    lon == "V8" ~ 8,
    lon == "V9" ~ 9,
    lon == "V10" ~ 10,
    lon == "V11" ~ 11,
    lon == "V12" ~ 12,
    lon == "V13" ~ 13,
    lon == "V14" ~ 14,
    lon == "V15" ~ 15,
    lon == "V16" ~ 16,
    lon == "V17" ~ 17,
    lon == "V18" ~ 18,
    lon == "V19" ~ 19,
    lon == "V20" ~ 20
  )) %>%
  mutate(year = as.numeric(year))


SSTdf_mean <- SSTdf_mean %>%
  mutate(mortality = 1 - (1 - exp((-(sst - 25)^2) / (4^2)))) %>% # caclulate mortaility and survival
  mutate(survival = ifelse(mortality > 0.7, 0.7, mortality)) %>%
  filter(lat %in% c(1, 50, 100)) %>%
  filter(lon == 10) %>%
  mutate(lat = as.factor(lat)) %>%
  mutate(lat = case_when(
    lat == 1 ~ "Lowest Latitude",
    lat == 100 ~ "Highest Latitude",
    lat == 50 ~ "Within MPA"
  )) %>% # rename latitutes
  filter(lat == "Within MPA") %>% # filter just 50
  mutate(climate = "Mean Shift") # name climate

SSTdf_enso <- SSTdf_enso %>%
  mutate(mortality = 1 - (1 - exp((-(sst - 25)^2) / (4^2)))) %>% # caclulate mortaility and survival
  mutate(survival = ifelse(mortality > 0.7, 0.7, mortality)) %>%
  filter(lat %in% c(1, 50, 100)) %>%
  filter(lon == 10) %>%
  mutate(lat = as.factor(lat)) %>%
  mutate(lat = case_when(
    lat == 1 ~ "Lowest Latitude",
    lat == 100 ~ "Highest Latitude",
    lat == 50 ~ "Within MPA"
  )) %>% # rename latitutes
  filter(lat == "Within MPA") %>% # filter just 50
  mutate(climate = "El Nino/ La Nina")

SSTdf_shock <- SSTdf_shock %>%
  mutate(mortality = 1 - (1 - exp((-(sst - 25)^2) / (4^2)))) %>% # caclulate mortaility and survival
  mutate(survival = ifelse(mortality > 0.7, 0.7, mortality)) %>%
  filter(lat %in% c(1, 50, 100)) %>%
  filter(lon == 10) %>%
  mutate(lat = as.factor(lat)) %>%
  mutate(lat = case_when(
    lat == 1 ~ "Lowest Latitude",
    lat == 100 ~ "Highest Latitude",
    lat == 50 ~ "Within MPA"
  )) %>% # rename latitutes
  filter(lat == "Within MPA") %>% # filter just 50
  mutate(climate = "Shocks")

SSTdf_mean_shock <- SSTdf_mean_shock %>%
  mutate(mortality = 1 - (1 - exp((-(sst - 25)^2) / (4^2)))) %>% # caclulate mortaility and survival
  mutate(survival = ifelse(mortality > 0.7, 0.7, mortality)) %>%
  filter(lat %in% c(1, 50, 100)) %>%
  filter(lon == 10) %>%
  mutate(lat = as.factor(lat)) %>%
  mutate(lat = case_when(
    lat == 1 ~ "Lowest Latitude",
    lat == 100 ~ "Highest Latitude",
    lat == 50 ~ "Within MPA"
  )) %>% # rename latitutes
  filter(lat == "Within MPA") %>% # filter just 50
  mutate(climate = "Shocks with Mean Shift")

SSTdf <- list(SSTdf_enso, SSTdf_mean, SSTdf_mean_shock, SSTdf_shock) %>%
  reduce(full_join) # combine all dataframes

temp <- ggplot(SSTdf) +
  geom_line(aes(year, sst), color = "blue") +
  geom_hline(aes(yintercept = 25), alpha = 0.5, linetype = "dashed") +
  geom_hline(aes(yintercept = 27), alpha = 0.5, linetype = "dashed", color = "red") +
  facet_wrap(~climate) +
  labs(x = "Time", y = "SST") +
  theme_bw() +
  scale_color_viridis_d() +
  theme(
    panel.grid = element_blank(),
    panel.grid.minor = element_blank(),
    strip.text = element_text(face = "bold"),
    strip.background = element_rect(fill = "white")
  ) + # plot temperatures
  scale_y_continuous(breaks = c(25, 27.5, 30, 32.5, 35), labels = c("0", "+ 2.5", "+ 5", "+ 7.5", "+ 10"))

ggsave(temp, file = paste0("climate_scenarios.pdf"), path = here::here("figs", "supp"), height = 10, width = 10)
```

```{r}
# Model Parameter set up --------------------------------------------------

SST <- seq(10, 40) # sea surface temperatures to model

opt.temp <- 25 # optimal temperature ranges

temp.range <- 4 # thermal performance range, w


# Mortality Function ------------------------------------------------------

calc_mortality <- function(SST, opt.temp, temp.range) {
  nat.m <- 1 - exp((-(SST - opt.temp)^2) / (temp.range^2)) # temperature based mortality function from Walsworth et al.
  return(1 - nat.m)
}


# Simulate Mortality Rates Function -----------------------------------------------

simulate_mortality <- function(SST, opt.temp, temp.range) {
  results <- matrix(NaN, nrow = length(SST), ncol = 1) # empty matrix to hold results


  for (i in 1:length(SST)) {
    for (j in 1:length(opt.temp)) {
      mort <- calc_mortality(SST[i], opt.temp, temp.range)
      results[i][j] <- mort
    }
  }

  results_plot <- as.data.frame(results) %>%
    rename(survival = V1) %>%
    mutate(survival.prop = case_when(
      survival > .7 ~ .7,
      survival < .7 ~ survival
    )) # caclulate mortality versus min ot temperature based and natural mortailtiy

  results_plot$SST <- SST
  results_plot$opt.temp <- opt.temp
  results_plot$range <- temp.range

  return(results_plot)
}

# Simulate multiple optimal temperatures ----------------------------------

simulation_results <- data.frame()

for (k in 1:length(opt.temp)) {
  sub_result <- simulate_mortality(SST, opt.temp[k], temp.range)
  simulation_results <- bind_rows(sub_result, simulation_results)
} # simulate across different optimum temperatures

simulation_results <- simulation_results %>%
  mutate(opt.temp = as.factor(opt.temp)) # make opt.temp a factor

# plot the mortality curve results
mort_curve <- ggplot(simulation_results, aes(SST, survival.prop)) +
  geom_line(linewidth = 1) +
  geom_vline(aes(xintercept = 25), linetype = "dashed", alpha = 0.5, color = "blue") +
  geom_vline(aes(xintercept = 27), linetype = "dotted", color = "blue") +
  geom_vline(aes(xintercept = 23), linetype = "dotted", color = "blue") +
  geom_vline(aes(xintercept = 34.19259), linetype = "longdash", alpha = 0.3, color = "red") +
  geom_vline(aes(xintercept = 31.84384), linetype = "longdash", alpha = 0.3, color = "red") +
  geom_vline(aes(xintercept = 30.93700), linetype = "longdash", alpha = 0.3, color = "red") +
  geom_vline(aes(xintercept = 29.66092), linetype = "longdash", alpha = 0.3, color = "red") +
  theme_bw() +
  labs(
    x = "Sea Surface Temperature",
    y = "Survival Probability"
  ) +
  theme(
    panel.grid.minor = element_blank(),
    legend.position = "none",
    axis.text.x = element_text(angle = 25, hjust = 1)
  ) +
  scale_x_continuous(breaks = c(23, 25, 27, 29.66092, 30.93700, 31.84384, 34.19259), labels = c("Thermal Min", "Thermal Optimum", "Thermal Max", "Shocks Max", "Mean Shift Max", "El Nino/ La Nina Max", "Shocks with Mean Shift Max"))

ggsave(mort_curve, file = paste0("mort_curve.pdf"), path = here::here("figs", "supp"), height = 10, width = 10)
```
